+++ galleria/plugins/__init__.py
@@ -1,5 +1,5 @@
 """Galleria plugin system."""
 
-from .base import BasePlugin
+from .base import BasePlugin, PluginContext, PluginResult
 
-__all__ = ["BasePlugin"]
+__all__ = ["BasePlugin", "PluginContext", "PluginResult"]